refactor(agents): extract shared stdin CLI response builder
The gemini and cursor CLI providers ran identically through runStdinCli and
duplicated the full ProviderResponse assembly: prompt telemetry construction,
parsed text/resolvedModel forwarding, and the exit-code -> providerError
cascade (127 binary_missing / 130 aborted / unavailable / non-zero retryable).
Only the binary name and failure label differed.

Factor that ~40-line block into buildStdinCliResponse() in stdin-cli-runner.ts
(its natural home, alongside runStdinCli). Each provider's generate() now ends
with a single call passing its binaryMissingMessage and clampFailure wrapper.
Parsers, command builders, env keys, and the _internals test surface are
unchanged, so behavior is identical.
