+++ packages/agents/src/providers/cursor-cli-provider.ts
@@ -1,7 +1,12 @@
 import type { ProcessManager } from '../process-manager';
-import { measurePromptPayload } from '../prompt-scope';
 import { firstString } from './output-summary';
-import { clampCliFailure, runStdinCli, type StdinCliCommand, stripAnsi } from './stdin-cli-runner';
+import {
+  buildStdinCliResponse,
+  clampCliFailure,
+  runStdinCli,
+  type StdinCliCommand,
+  stripAnsi,
+} from './stdin-cli-runner';
 import type { AgentProvider, ProviderPhase, ProviderRequest, ProviderResponse } from './types';
 
 const CURSOR_ENV_KEYS = [
@@ -107,46 +112,10 @@ export function createCursorCliProvider(processManager: ProcessManager): AgentPr
         lifecycleMessage: 'Cursor CLI started',
         unavailableMessage: 'Cursor CLI stdin execution is unavailable',
       });
-      const parsed = parseCursorOutput(result.rawOutput);
-      const promptTelemetry = {
-        phase: req.phase,
-        promptSize: measurePromptPayload(req.prompt),
-        ...(req.promptMaterialSummary ? { selectedMaterials: req.promptMaterialSummary } : {}),
-      };
-
-      return {
-        rawOutput: parsed.text,
-        exitCode: result.exitCode,
-        promptTelemetry,
-        ...(parsed.resolvedModel ? { resolvedModel: parsed.resolvedModel } : {}),
-        ...(result.exitCode === 127
-          ? {
-              providerError: {
-                kind: 'binary_missing' as const,
-                message: 'cursor-agent CLI not found on PATH',
-                retryable: false,
-              },
-            }
-          : result.exitCode === 130
-            ? { providerError: { kind: 'aborted' as const, message: 'aborted', retryable: false } }
-            : result.unavailable
-              ? {
-                  providerError: {
-                    kind: 'unknown' as const,
-                    message: clampCursorFailure(result.rawOutput, req.prompt),
-                    retryable: false,
-                  },
-                }
-              : result.exitCode !== 0
-                ? {
-                    providerError: {
-                      kind: 'unknown' as const,
-                      message: clampCursorFailure(result.rawOutput, req.prompt),
-                      retryable: true,
-                    },
-                  }
-                : {}),
-      };
+      return buildStdinCliResponse(req, result, parseCursorOutput(result.rawOutput), {
+        binaryMissingMessage: 'cursor-agent CLI not found on PATH',
+        clampFailure: clampCursorFailure,
+      });
     },
     async healthCheck() {
       return { ok: true };
